Purge the imagej-legacy dependency
We can check whether the application frame is a UIComponent,
then check whether its associated component is a Frame,
without having a hard dependency on the LegacyApplicationFrame.
This reduces n5-ij's dependency footprint, and avoids downstream
issues with the too-late patching of net.imagej:ij that can occur
in some scenarios when imagej-legacy is on the classpath.

This also reduces the dependency count from 180 to 135.